feat: Implement field and field group classes with definitions for angle and object height fields
