+++ optiland/fields/__init__.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from .field import Field
+from .field_group import FieldGroup
+from .field_types import AngleField, BaseFieldDefinition, ObjectHeightField
+
+__all__ = [
+    "Field",
+    "FieldGroup",
+    "BaseFieldDefinition",
+    "AngleField",
+    "ObjectHeightField",
+]
